Added an author to the entry.
This is mainly to satisfy the Atom specification, that states that if the
entry is to be used as a stand-alone document (which it may be in this
implementation), it MUST have at least one author, and the author MUST
have at least a name.
At the moment I've simply hard-coded the name to "Grean", but in a more
advanced scenario, we could consider pulling the name from a set of
claims, so that we actually have a sort of audit trail of the originators
of each changeset.
For that reason, I also made the test code quite lax in its requirements.