Added UuidIri struct,
intended to be a Value Object modeling a uuid IRI. The motivation is that
I've already identifed three places in the code base where it refers to
the hard-coded string "urn:uuid", which tells me that I can invoke the
Rule of Three. I foresee more uuid parsing and matching in my immediate
future.