Added an Author to the feed.
This is a semi-mandatory field, in the sense that it MUST appear on either
Atom feed or on each Atom entry, so for good measure this guarantees that
there's also an Author on the feed itself.
As with the Atom entry, the Author is currently hard-coded to "Grean", but
we should consider using an authentication claim in order to audit who
caused the changeset.
There's currently a bit of duplication in the code, but each particular
piece of code only appears twice, so I'm invoking the Rule of Three.