Added a Parse method to UuidIri.
Notice that the existing TryParse tests have been repurposed. This breaks
with the standard AAA/Four-Phase Test pattern, but the rationale for
breaking the pattern is as follows:
When testing a method pair like Parse/TryParse, you're basically testing
the same behavior. There's just two different ways of expressing the same
behavior.
This is evident because for both methods, the same test cases should be
applied, with exactly the same Fixture Setup.
This is a new principle that I'm currently trying on for size. It can also
be applied to:
- IEnumerable.GetEnumerator()/IEnumerable<T>.GetEnumerator()
- Object.Equals(object)/IEquatable<T>.Equals(T)
- IComparable.CompareTo(object)/IComparable<T>.CompareTo(T)
There are probably even more of such pairs in various idioms and
interfaces in the BCL...