Encapsulated change of rel types in an ext.method,
enabling a Fluent style of expressing the intent of the test.
Although the ChangeLinkRelationShipTypes extension method also performs a
clone, I decided to keep the explicit call to Clone in the test method so
that it's more apparent that this is a new object, and not the original
being modified.