Implemented ISyndicationFeedReader
on AtomFileAccess.
Despite the Cyclomatic Complexity of 2+ in the Read method, I prefer
keeping this a Humble Object. The alternative is either to attempt to set
up and be sure to tear down files in the file system, which when it comes
to automated testing can be error prone; or to introduce yet another
abstraction that can check if the file already exists and create an
XmlReader instance if it does. I don't think the advantage is worth the
added complexity.