Renamed Read method to ReadFeed,
because I'll soon introduce a Read method that returns a SyndicationItem
instead of a SyndicationFeed, and if both are named "Read" at least one of
them would have to be explicit interface implementations. One could argue
that this reasoning is a Leaky Abstraction, but I find ReadFeed to be a
good, explicit name too; the name itself betrays no leakiness.