Made SyndicationEventStream generic,
which doesn't seem to provide a lot of benefit at the moment, but should
start to make more sense once it gets turned into an IEnumerable<T> as
well.
Some consumers may still prefer to treat events as completely untyped, and
they can still do that by declaring a SyndicationEventStream<object>, but
other consumers might e.g. want to declare all their events as a
participant in the Visitor pattern. Making the SyndicationStream generic
enables the client to declare a SyndicationStream<EventElement>, which
will make reading the stream easier, and type-safe.