Used XmlSerializer instd of DataContractSerializer
because when serializing/deserializing Atom (or RSS, for that matter), we
are already explicitly dealing with XML, and we might as well be explicit
about this.
There are things that the default DataContractSerializer can't do, such as
handle XML attributes, and there's no reason to limit ourselves from using
XML attributes. Thus, I think that the default serializer should be an
XmlSerializer.