Added a WithName method.
This may seem totally redundant at the moment, but if the Value Object is
expanded to also contain its other, optional properties, this begins to
make sense.
This is also the reason that the test verification is a bit more
convoluted than it has to be, because the likeness ensures that any other
values are kept equal.