Added WithContent.
This marks the last of the WithXyz methods. Why no WithId or WithPublished
methods? Basically, these WithXyz methods are the immutable equivalent of
mutating the object. You take an existing instance and change a single
property at the time. However, and Atom entry is identied by its ID, so if
you change the ID, you essentially create a completely new entry. The API
for that should be more explicit, because I don't think there's a good
case for that... maybe there is, but then we can add the method.
The same sort of argument applies to the Published property.