Added AtomEventsOnAzure class,
which will implement the Azure Blob Storage-specifics for AtomEventStore.
Due to the nature of this class, the only way to test it would be against
the Azure Blob Storage itself, which would indicate a slow and fragile set
of integration tests. It's my hope that the implementation is going to be
simple enough that we can treat it as a Humble Object, so that's what I'll
do until proven wrong about that assumption.