Incremented minor version,
in order to indicate that the introduction of the IContentSerializer and
the XmlContentSerializer class is most definitely a new feature. Actually,
the addition of IContentSerializer to the constructor of AtomEventStream
is a breaking change, but according to Semantic Versioning 2.0.0, since
AtomEventStore is still in the major version 0 range, anything can change
without further notice.