Changed the addressing scheme to be more indexable
The adressing scheme is now that each event stream consists of two URL
segments, the first of which is the ID of the event stream, and the next
is the ID of the Feed page itself.

For example, the event stream with the ID
10B1F47A-154F-49C7-B553-42691831F452 could have the following Feed pages:

10B1F47A-154F-49C7-B553-42691831F452/10B1F47A-154F-49C7-B553-42691831F452
10B1F47A-154F-49C7-B553-42691831F452/A4D040FF-25A1-43E3-96E6-28AC21408B18
10B1F47A-154F-49C7-B553-42691831F452/692BE599-50A1-4B63-9883-61C367C3B509
10B1F47A-154F-49C7-B553-42691831F452/A0CAE02B-80ED-4C34-916B-D9F02E89453E

Notice that the first segment is the same value for all four pages: the ID
of the event stream. Also notice that the first of these four addresses
repeat the ID in the second segment. This is the index of the event
stream.