Skipped reading forward until second element,
because it's a fairly common to check to see if the stream is empty, or to
get the first element, and such operations shouldn't trigger a
read-forward. The assumption here is that if a client reads past the first
element, there's a good chance that it will want to read more events. Not
guaranteed, but more likely.