Refactored a test to use DataContractSerializer,
instead of ConventionBasedSerializerOfComplexImmutableClasses, in order to
decouple the tests from this legacy class. The concern of this particular
test case was to test that an AtomEntry is correctly rendered as XML. The
content rendering isn't the primary target of this test, but rather the
metadata (Id, Title, Author, etc.) is.